requirements_api.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.4
slowapi==0.1.9
websockets==12.0
bcrypt==4.1.2
